+++ node/src/main.rs
@@ -1,7 +1,11 @@
-//! Substrate Node Template CLI library.
+//! Substrate Parachain Node Template CLI
+
 #![warn(missing_docs)]
 
-mod benchmarking;
+use polkadot_sdk::*;
+
+// TODO: has some compilation problems and disabled now
+// mod benchmarking;
 mod chain_spec;
 mod cli;
 mod command;
@@ -9,5 +13,5 @@ mod rpc;
 mod service;
 
 fn main() -> sc_cli::Result<()> {
-    command::run()
+	command::run()
 }
