--- parachain/node/src/chain_spec.rs
@@ -1,167 +1,0 @@
-use polkadot_sdk::*;
-
-use qf_parachain_runtime as runtime;
-use qf_parachain_runtime::{AccountId, Signature, WASM_BINARY};
-use sc_chain_spec::{ChainSpecExtension, ChainSpecGroup};
-use qfp_consensus_spin::sr25519::AuthorityId as SpinId;
-use sc_service::ChainType;
-use sp_consensus_grandpa::AuthorityId as GrandpaId;
-use sp_core::{Pair, Public, sr25519};
-use sp_runtime::traits::{IdentifyAccount, Verify};
-use serde::{Deserialize, Serialize};
-
-/// Specialized `ChainSpec` for the normal parachain runtime.
-pub type ChainSpec = sc_service::GenericChainSpec<Extensions>;
-
-/// Generate a crypto pair from seed.
-pub fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
-    TPublic::Pair::from_string(&format!("//{}", seed), None)
-        .expect("static values are valid; qed")
-        .public()
-}
-
-type AccountPublic = <Signature as Verify>::Signer;
-
-/// Generate an account ID from seed.
-pub fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
-where
-    AccountPublic: From<<TPublic::Pair as Pair>::Public>,
-{
-    AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
-}
-
-/// Generate an SPIN authority key.
-pub fn authority_keys_from_seed(s: &str) -> (SpinId, GrandpaId) {
-    (get_from_seed::<SpinId>(s), get_from_seed::<GrandpaId>(s))
-}
-
-/// The relay chain that you want to configure this parachain to connect to.
-pub const RELAY_CHAIN: &str = "rococo-local";
-
-/// The extensions for the [`ChainSpec`].
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ChainSpecGroup, ChainSpecExtension)]
-pub struct Extensions {
-	/// The relay chain of the Parachain.
-	#[serde(alias = "relayChain", alias = "RelayChain")]
-	pub relay_chain: String,
-	/// The id of the Parachain.
-	#[serde(alias = "paraId", alias = "ParaId")]
-	pub para_id: u32,
-}
-
-impl Extensions {
-	/// Try to get the extension from the given `ChainSpec`.
-	pub fn try_get(chain_spec: &dyn sc_service::ChainSpec) -> Option<&Self> {
-		sc_chain_spec::get_extension(chain_spec.extensions())
-	}
-}
-
-/// TODO: not compile for SpinId
-/// Configure initial storage state for FRAME modules.
-fn testnet_genesis(
-    initial_authorities: Vec<(SpinId, GrandpaId)>,
-    root_key: AccountId,
-    endowed_accounts: Vec<AccountId>,
-    _enable_println: bool,
-) -> serde_json::Value {
-    serde_json::json!({
-        "balances": {
-            // Configure endowed accounts with initial balance of 1 << 63.
-            "balances": endowed_accounts.iter().cloned().map(|k| (k, 1u64 << 63)).collect::<Vec<_>>(),
-        },
-        "aura": {
-            "authorities": initial_authorities.iter().map(|x| (x.0.clone())).collect::<Vec<_>>(),
-        },
-        "grandpa": {
-            "authorities": initial_authorities.iter().map(|x| (x.1.clone(), 1)).collect::<Vec<_>>(),
-        },
-        "sudo": {
-            // Assign network admin rights.
-            "key": Some(root_key),
-        },
-    })
-}
-
-/// NOTE: file not found
-pub fn qf_devnet_config() -> Result<ChainSpec, String> {
-    ChainSpec::from_json_bytes(&include_bytes!("./res/qf-devnet.raw.json")[..])
-}
-
-pub fn development_chain_spec() -> ChainSpec {
-	// Give your base currency a unit name and decimal places
-	let mut properties = sc_chain_spec::Properties::new();
-	properties.insert("tokenSymbol".into(), "UNIT".into());
-	properties.insert("tokenDecimals".into(), 12.into());
-	properties.insert("ss58Format".into(), 42.into());
-
-	ChainSpec::builder(
-		runtime::WASM_BINARY.expect("WASM binary was not built, please build it!"),
-		Extensions { relay_chain: RELAY_CHAIN.into(), para_id: runtime::PARACHAIN_ID },
-	)
-	.with_name("Development")
-	.with_id("dev")
-	.with_chain_type(ChainType::Development)
-	.with_genesis_config_preset_name(sp_genesis_builder::DEV_RUNTIME_PRESET)
-	.with_properties(properties)
-    .with_genesis_config_patch(testnet_genesis(
-        // Initial PoA authorities
-        vec![authority_keys_from_seed("Alice")],
-        // Sudo account
-        get_account_id_from_seed::<sr25519::Public>("Alice"),
-        // Pre-funded accounts
-        vec![
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            get_account_id_from_seed::<sr25519::Public>("Bob"),
-            get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
-            get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
-        ],
-        true,
-    ))
-	.build()
-}
-
-pub fn local_chain_spec() -> ChainSpec {
-	// Give your base currency a unit name and decimal places
-	let mut properties = sc_chain_spec::Properties::new();
-	properties.insert("tokenSymbol".into(), "UNIT".into());
-	properties.insert("tokenDecimals".into(), 12.into());
-	properties.insert("ss58Format".into(), 42.into());
-
-	#[allow(deprecated)]
-	ChainSpec::builder(
-		runtime::WASM_BINARY.expect("WASM binary was not built, please build it!"),
-		Extensions { relay_chain: RELAY_CHAIN.into(), para_id: runtime::PARACHAIN_ID },
-	)
-	.with_name("Local Testnet")
-	.with_id("local_testnet")
-	.with_chain_type(ChainType::Local)
-	.with_genesis_config_preset_name(sc_chain_spec::LOCAL_TESTNET_RUNTIME_PRESET)
-	.with_protocol_id("template-local")
-	.with_properties(properties)
-    .with_genesis_config_patch(testnet_genesis(
-        // Initial PoA authorities
-        vec![
-            authority_keys_from_seed("Alice"),
-            authority_keys_from_seed("Bob"),
-        ],
-        // Sudo account
-        get_account_id_from_seed::<sr25519::Public>("Alice"),
-        // Pre-funded accounts
-        vec![
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            get_account_id_from_seed::<sr25519::Public>("Bob"),
-            get_account_id_from_seed::<sr25519::Public>("Charlie"),
-            get_account_id_from_seed::<sr25519::Public>("Dave"),
-            get_account_id_from_seed::<sr25519::Public>("Eve"),
-            get_account_id_from_seed::<sr25519::Public>("Ferdie"),
-            get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
-            get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
-            get_account_id_from_seed::<sr25519::Public>("Charlie//stash"),
-            get_account_id_from_seed::<sr25519::Public>("Dave//stash"),
-            get_account_id_from_seed::<sr25519::Public>("Eve//stash"),
-            get_account_id_from_seed::<sr25519::Public>("Ferdie//stash"),
-        ],
-        true,
-    ))
-	.build()
-}
