--- parachain/node/src/main.rs
@@ -1,16 +1,0 @@
-//! Substrate Parachain Node Template CLI
-
-#![warn(missing_docs)]
-
-use polkadot_sdk::*;
-
-// mod benchmarking;
-mod chain_spec;
-mod cli;
-mod command;
-mod rpc;
-mod service;
-
-fn main() -> sc_cli::Result<()> {
-	command::run()
-}
