--- runtimes/parachain/Cargo.toml
@@ -1,215 +1,0 @@
-[package]
-name = "qf-parachain-runtime"
-description = "A parachain runtime built with Substrate and Cumulus."
-version = "0.1.0"
-license.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-edition.workspace = true
-publish = false
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-[dependencies]
-codec = { features = ["derive"], workspace = true }
-docify = { workspace = true }
-hex-literal = { optional = true, workspace = true, default-features = true }
-log = { workspace = true }
-scale-info = { features = ["derive", "serde"], workspace = true }
-serde_json = { workspace = true, default-features = false, features = ["alloc"] }
-smallvec = { workspace = true, default-features = true }
-
-polkadot-sdk = { workspace = true, default-features = false, features = [
-	"pallet-aura",
-	"pallet-authorship",
-	"pallet-balances",
-	"pallet-message-queue",
-	"pallet-session",
-	"pallet-sudo",
-	"pallet-timestamp",
-	"pallet-transaction-payment",
-	"pallet-transaction-payment-rpc-runtime-api",
-
-	"pallet-xcm",
-	"polkadot-parachain-primitives",
-	"polkadot-runtime-common",
-	"staging-xcm",
-	"staging-xcm-builder",
-	"staging-xcm-executor",
-
-	"cumulus-pallet-aura-ext",
-	"cumulus-pallet-session-benchmarking",
-	"cumulus-pallet-weight-reclaim",
-	"cumulus-pallet-xcm",
-	"cumulus-pallet-xcmp-queue",
-	"cumulus-primitives-aura",
-	"cumulus-primitives-core",
-	"cumulus-primitives-utility",
-	"pallet-collator-selection",
-	"parachains-common",
-	"staging-parachain-info",
-
-	"runtime",
-] }
-
-# Cumulus
-cumulus-pallet-parachain-system = { workspace = true }
-
-
-# frame
-frame-support = { features = ["experimental"], workspace = true }
-frame-system = { workspace = true }
-frame-try-runtime = { optional = true, workspace = true }
-frame-executive = { workspace = true }
-frame-metadata-hash-extension = { workspace = true }
-
-# frame pallets
-pallet-aura = { workspace = true }
-pallet-balances = { workspace = true }
-pallet-grandpa = { workspace = true }
-pallet-sudo = { workspace = true }
-pallet-timestamp = { workspace = true }
-pallet-transaction-payment = { workspace = true }
-
-# QF pallets
-pallet-faucet = {workspace = true}
-pallet-qf-polkavm = { workspace = true }
-pallet-qf-polkavm-dev = { workspace = true }
-
-# primitives
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-qfp-consensus-spin = { features = [
-	"serde",
-], workspace = true }
-sp-consensus-grandpa = { features = [
-	"serde",
-], workspace = true }
-sp-core = { features = [
-	"serde",
-], workspace = true }
-sp-inherents = { workspace = true }
-sp-offchain = { workspace = true }
-sp-runtime = { features = [
-	"serde",
-], workspace = true }
-sp-session = { workspace = true }
-sp-storage = { workspace = true }
-sp-transaction-pool = { workspace = true }
-sp-version = { features = [
-	"serde",
-], workspace = true }
-sp-genesis-builder = { workspace = true }
-
-# RPC related
-frame-system-rpc-runtime-api = { workspace = true }
-pallet-transaction-payment-rpc-runtime-api = { workspace = true }
-
-# Used for runtime benchmarking
-frame-benchmarking = { optional = true, workspace = true }
-frame-system-benchmarking = { optional = true, workspace = true }
-
-# The pallet in this template.
-# pallet-template = { workspace = true }
-
-[build-dependencies]
-docify = { workspace = true }
-substrate-wasm-builder = { optional = true, workspace = true, default-features = true }
-
-
-
-[features]
-default = ["std"]
-std = [
-	"codec/std",
-	"cumulus-pallet-parachain-system/std",
-	"log/std",
-	"polkadot-sdk/std",
-	"scale-info/std",
-	"serde_json/std",
-
-	"frame-executive/std",
-	"frame-metadata-hash-extension/std",
-	"frame-support/std",
-	"frame-system-benchmarking?/std",
-	"frame-system-rpc-runtime-api/std",
-	"frame-system/std",
-
-	"frame-benchmarking?/std",
-	"frame-try-runtime?/std",
-
-	"pallet-aura/std",
-	"pallet-balances/std",
-	"pallet-grandpa/std",
-	"pallet-sudo/std",
-	"pallet-qf-polkavm/std",
-	"pallet-qf-polkavm-dev/std",
-	"pallet-timestamp/std",
-	"pallet-transaction-payment-rpc-runtime-api/std",
-	"pallet-transaction-payment/std",
-
-	"sp-api/std",
-	"sp-block-builder/std",
-	"qfp-consensus-spin/std",
-	"sp-consensus-grandpa/std",
-	"sp-core/std",
-	"sp-genesis-builder/std",
-	"sp-inherents/std",
-	"sp-offchain/std",
-	"sp-runtime/std",
-	"sp-session/std",
-	"sp-storage/std",
-	"sp-transaction-pool/std",
-	"sp-version/std",
-
-	"substrate-wasm-builder",
-]
-
-runtime-benchmarks = [
-	"cumulus-pallet-parachain-system/runtime-benchmarks",
-	"hex-literal",
-	"polkadot-sdk/runtime-benchmarks",
-	"frame-benchmarking/runtime-benchmarks",
-	"frame-support/runtime-benchmarks",
-	"frame-system-benchmarking/runtime-benchmarks",
-	"frame-system/runtime-benchmarks",
-	"pallet-balances/runtime-benchmarks",
-	"pallet-grandpa/runtime-benchmarks",
-	"pallet-sudo/runtime-benchmarks",
-	"pallet-qf-polkavm/runtime-benchmarks",
-	"pallet-qf-polkavm-dev/runtime-benchmarks",
-	"pallet-timestamp/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks",
-]
-
-try-runtime = [
-	"cumulus-pallet-parachain-system/try-runtime",
-	"polkadot-sdk/try-runtime",
-	"frame-executive/try-runtime",
-	"frame-support/try-runtime",
-	"frame-system/try-runtime",
-	"frame-try-runtime/try-runtime",
-	"pallet-aura/try-runtime",
-	"pallet-balances/try-runtime",
-	"pallet-grandpa/try-runtime",
-	"pallet-sudo/try-runtime",
-	"pallet-qf-polkavm/try-runtime",
-	"pallet-qf-polkavm-dev/try-runtime",
-	"pallet-timestamp/try-runtime",
-	"pallet-transaction-payment/try-runtime",
-	"sp-runtime/try-runtime",
-]
-
-# Enable the metadata hash generation.
-#
-# This is hidden behind a feature because it increases the compile time.
-# The wasm binary needs to be compiled twice, once to fetch the metadata,
-# generate the metadata hash and then a second time with the
-# `RUNTIME_METADATA_HASH` environment variable set for the `CheckMetadataHash`
-# extension.
-metadata-hash = ["substrate-wasm-builder/metadata-hash"]
-
-# A convenience feature for enabling things when doing a build
-# for an on-chain release.
-on-chain-release-build = ["metadata-hash", "sp-api/disable-logging"]
