--- runtimes/parachain/README.md
@@ -1,10 +1,0 @@
-# Runtime
-
-ℹ️ The runtime (in other words, a state transition function), refers to the core logic of the parachain that is
-responsible for validating blocks and executing the state changes they define.
-
-💁 The runtime in this template is constructed using ready-made FRAME pallets that ship with
-[Polkadot SDK](https://github.com/paritytech/polkadot-sdk), and a [template for a custom pallet](../pallets/README.md).
-
-👉 Learn more about FRAME
-[here](https://paritytech.github.io/polkadot-sdk/master/polkadot_sdk_docs/polkadot_sdk/frame_runtime/index.html).
